Agrega a la tabla carrito listo

+++ orders/models.py
@@ -11,28 +11,28 @@ class Toppings (models.Model):
 class DinnerPlates (models.Model):
     name_dinner = models.CharField(max_length=40, blank=True)
     tamanio = models.CharField(max_length=10, blank=True)
-    price_dinner = models.DecimalField(max_digits=8, decimal_places=4, blank=True)
+    price_dinner = models.FloatField(  blank=True)
 
     def __str__(self):
         return f"{self.nombre_tooping} - {self.tamanio} - {self.price_dinner}"
 
 class Pasta (models.Model):
     name_pasta = models.CharField(max_length=40, blank=True)
-    price_pasta = models.DecimalField(max_digits=8, decimal_places=6, blank=True)
+    price_pasta = models.FloatField(  blank=True)
     
     def __str__(self):
         return f"{self.name_pasta} - {self.price_pasta}"
 
 class Salads (models.Model):
     name_salads = models.CharField(max_length=40, blank=True)
-    price_salads = models.DecimalField(max_digits=5, decimal_places=4, blank=True)
+    price_salads = models.FloatField(  blank=True)
 
     def __str__(self):
         return f"{self.name_salads} - {self.price_salads}"
 
 class ExtraSubs(models.Model):
     name_extras = models.CharField(max_length=25, blank=True)
-    price_extra_sub = models.DecimalField(max_digits=5, decimal_places=4, blank=True)
+    price_extra_sub = models.FloatField(  blank=True)
 
     def __str__(self):
         return f"{self.name_extras} - {self.price_extra_sub}"
@@ -41,7 +41,7 @@ class Subs (models.Model):
     #extra_sub = models.ManyToManyField(ExtraSubs)
     name_subs = models.CharField(max_length=40, blank=True)
     tamanio = models.CharField(max_length=10, blank=True)
-    price_subs = models.DecimalField(max_digits=5, decimal_places=2, blank=True)
+    price_subs = models.FloatField(  blank=True)
 
     def __str__(self):
         return f"{self.name_subs} - {self.tamanio} - {self.price_subs}"
@@ -49,7 +49,7 @@ class Subs (models.Model):
 class RegularPizza(models.Model):
     name_reg_pizza = models.CharField(max_length=30, blank=True)
     tamanio = models.CharField(max_length=10, blank=True)
-    price_reg_pizza = models.DecimalField(max_digits=5, decimal_places=2, blank=True)
+    price_reg_pizza = models.FloatField(  blank=True)
     quantity_top = models.IntegerField(max_length=1, blank=True)
     #toppings = models.ManyToManyField(Toppings)
 
@@ -59,7 +59,7 @@ class RegularPizza(models.Model):
 class SicilianPizza(models.Model):
     name_sic_pizza = models.CharField(max_length=30, blank=True)
     tamanio = models.CharField(max_length=10, blank=True)
-    price_sic_pizza = models.DecimalField(max_digits=5, decimal_places=2, blank=True)
+    price_sic_pizza = models.FloatField(  blank=True)
     quantity_top = models.IntegerField(max_length=1, blank=True)
     #toppings = models.ManyToManyField(Toppings)
 
@@ -69,7 +69,7 @@ class SicilianPizza(models.Model):
 class SpecialRegularPizza(models.Model):
     name_special_reg = models.CharField(max_length=40 ,default="Special Regular Pizza")
     tamanio = models.CharField(max_length=10, blank=True)
-    price_special_reg = models.DecimalField(max_digits=5, decimal_places=2, blank=True)
+    price_special_reg = models.FloatField(  blank=True)
     #toppings = models.ManyToManyField(Toppings)
 
     def __str__(self):
@@ -78,7 +78,7 @@ class SpecialRegularPizza(models.Model):
 class SpecialSicilianPizza(models.Model):
     name_special_sic = models.CharField(max_length=40 ,default="Special Sicilian Pizza")
     tamanio = models.CharField(max_length=10, blank=True)
-    price_special_sic = models.DecimalField(max_digits=5, decimal_places=2,blank=True)
+    price_special_sic = models.FloatField( blank=True)
     #toppings = models.ManyToManyField(Toppings)
 
     def __str__(self):
@@ -88,7 +88,7 @@ class Carrito(models.Model):
     status = models.BooleanField(default=False)
     items = models.CharField(max_length=200, blank=True)
     cantidad = models.IntegerField(default=0, blank=True)
-    price_items = models.DecimalField(max_digits=6, decimal_places=2, blank=True)
+    price_items = models.FloatField(  blank=True)
     fecha_items = models.DateField(auto_now_add=True)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
